% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_influence.R
\name{calculate_influence}
\alias{calculate_influence}
\title{calculate_influence}
\usage{
calculate_influence(conf_obj, data_df)
}
\arguments{
\item{conf_obj}{object for which the obj element has been optimsied}

\item{data_df}{data.frame of original data set}
}
\value{
list objects that summarise of influence statistics that can then be passed to a plotting function
}
\description{
calculate_influence
}
\details{
given an optimsied model that has been output from configure_obj() function, we will generate a bunch of influence statistics following the work from Nokome's influence plot metrics.
Bentley et.al. 2011 Influence plots and metrics: tools for better understanding fisheries catch-per-unit-effort standardizations
}
