% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{canonical_fun}
\alias{canonical_fun}
\title{canonical_fun}
\usage{
canonical_fun(mu, dist, k = NULL)
}
\arguments{
\item{mu}{fitted values from the model}

\item{dist}{character values options availble {gaussian, gamma,binomial, neg_binomial, inverse_gaussian, poisson}}

\item{k}{the k parameter needed for the negative binomial evaluation}
}
\value{
canonical given fitted value
}
\description{
canonical_fun
}
\details{
returns the canonical form (often denoted as theta in the literature) given mu
}
\keyword{internal}
