% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{check_dims}
\alias{check_dims}
\title{check_dims}
\usage{
check_dims(m1, m2, row)
}
\arguments{
\item{m1}{a matrix}

\item{m2}{a matrix}

\item{row}{boolean check row if false check columns}
}
\value{
list with bool and message
}
\description{
check_dims
}
\details{
check a dimension is consistent between two matricies
}
\keyword{internal}
