% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_inputs}
\alias{check_inputs}
\title{check_inputs}
\usage{
check_inputs(tmb_data, tmb_params)
}
\arguments{
\item{tmb_data}{a list of data}

\item{tmb_params}{a list of parameters}
}
\value{
list with two elements a boolean called result (True if everything is okay and False if flagged warnigns) and string vector of errors if result == FALSE
}
\description{
check_inputs
}
\details{
a wrapper function to check data is compatible with parameters, so we don't get a segment fault from the TMB MakeAdFun
}
\keyword{internal}
