% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPUEspatial-package.R
\docType{package}
\name{CPUEspatial-package}
\alias{CPUEspatial}
\alias{CPUEspatial-package}
\title{CPUEspatial: Geostatistical CPUE analysis for univariate response variables.}
\description{
Geostatistical CPUE analysis, currently only univariate, plans to allow for hurdle models.
}
\author{
\strong{Maintainer}: Craig Marsh \email{craig.marsh10@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
