% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{cumulant_fun}
\alias{cumulant_fun}
\title{cumulant_fun}
\usage{
cumulant_fun(theta, dist)
}
\arguments{
\item{theta}{canonical form of the distirbution}

\item{dist}{character values options availble {gaussian, gamma,binomial, neg_binomial, inverse_gaussian, poisson}}
}
\value{
returns the evaluation of the cumulative function given a theta
}
\description{
cumulant_fun
}
\details{
This is the cumulative function for the exponential families table 5.1 of GLM with examples in R DUnn 2018
}
\keyword{internal}
