% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{deviance_calc}
\alias{deviance_calc}
\title{deviance_calc}
\usage{
deviance_calc(y, mu, dist, phi = NULL)
}
\arguments{
\item{y}{response variable values}

\item{mu}{fitted values from the model}

\item{dist}{character values options availble {gaussian, gamma,binomial, neg_binomial, inverse_gaussian, poisson}}

\item{k}{the k parameter needed for the negative binomial evaluation}
}
\value{
residual deviance
}
\description{
deviance_calc
}
\details{
returns the residual deviance based on the exponential family. That is 2 time fitted compared with the saturated model
}
