% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{evalit}
\alias{evalit}
\title{evalit}
\usage{
evalit(x, ...)
}
\arguments{
\item{x}{a string to}
}
\value{
what ever the evaluation string defines
}
\description{
evalit
}
\details{
evaluates a character string as if you typed it into the console
}
\examples{
evalit("print('hello')")
df = data.frame(x = c(1,32,4,3), y = c(23,32,1,1))
var_label = "x"
evalit(paste0("df$",var_label)) # should return df$x
}
\keyword{internal}
