% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_and_merge.R
\name{extract_and_merge}
\alias{extract_and_merge}
\title{extract_and_merge}
\usage{
extract_and_merge(par_label, reports, report_labels)
}
\arguments{
\item{par_label}{character string of the parameter you want to extract}

\item{reports}{A list of lists. Each element is obj$report(), this will allow for multiple models}

\item{report_labels}{when creating the the data frame this will be the label representing each element in the list}
}
\value{
data.frame of quantities for each model
}
\description{
extract_and_merge
}
\details{
an function for pulling out parameters from a list object that has been created from obj$report(). Used to create a data.frame
with multiple models.
}
\examples{
\dontrun{
 n_sims = 10
 sim_ls = list()
 for(i in 1:n_sims)
   sim_ls[[i]] = obj$report()
 reports = list(sim_ls)  
 extract_and_merge("sigma", reports, report_labels = "sim")  
}

}
