% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correlation.R
\name{get_correlation}
\alias{get_correlation}
\title{get_correlation}
\usage{
get_correlation(
  observed_df,
  projection_df,
  time_variable_label,
  proj_variable_label,
  projection_raster_layer
)
}
\arguments{
\item{observed_df}{SpatialPointsDataFrame, which contains response variable and covariates for glmm analysis mut contain column 'area'}

\item{projection_df}{SpatialPointsDataFrameneeds to have the same variable names (colnames) as observed_df. Should supply variable values for all projection cells over all time steps}

\item{time_variable_label}{character relating to a column names needs to be an integer variable}

\item{proj_variable_label}{character relating to a column name for the projected variable, usually derived from get_projection()}

\item{projection_raster_layer}{a RasterLayer object only required if apply_preferential_sampling = TRUE, and preference_model_type == 1. Should be the same resolution as projection_df. Used to collate sample locations. The observed_df slot should have vaules 0 = cell not in projection grid or 1 = active projection cell}
}
\value{
: a list with the following elements
\itemize{
  \item samples_rasters list of raster layers for each time-step with the accumulated observations in each projection grid
  \item proj_rasters list of raster layers for each time-step with the projected value
  \item correlation_by_time_step Pearsons correlation between complete obs of the two raster layers by time-step. non-sampled areas not included
  \item correlation_by_time_step_alt Pearsons correlation where non-sampled cells are set = 0 and used in correlation.
  \item overall_correlation Pearsons correlation for complete obs over all time steps
}
}
\description{
get_correlation
}
\details{
Calculate the correlation between number of samples in a projected cell with predicted spatial abundance from a model
with no Preferential sampling. Used as an exploratory tool to see if Preferential sampling could be a useful model addition
}
\examples{
\dontrun{
you can plot the layers side by side using the following R-code
corr_obj = get_correlation(...)
plot for the first time-step
par(mfrow = c(2,1))
plot(corr_obj$samples_rasters[[1]])
plot(proj_rasters$samples_rasters[[1]])
}
}
