% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projection.R
\name{get_projection}
\alias{get_projection}
\title{get_projection}
\usage{
get_projection(obj, data, projection_df, time_variable_label)
}
\arguments{
\item{obj}{TMB object which has been optimised and checkef that it successfully converged}

\item{data}{number of simulated data sets}

\item{projection_df}{spatial data frame}

\item{time_variable_label}{character of}
}
\value{
: projection_df with a column called predicted_y which has all the spatial components plus the time_coeffecient
}
\description{
get_projection
}
\details{

}
