% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{inverse_link}
\alias{inverse_link}
\title{inverse_link}
\usage{
inverse_link(x, link_fun)
}
\arguments{
\item{x}{a vector of values to transform}

\item{link_fun}{integer based on the above}
}
\value{

}
\description{
inverse_link
}
\details{
applies the inverse of the link function
log_link                 = 0
logit_link               = 1
probit_link              = 2
inverse_link             = 3
identity_link            = 4
}
\keyword{internal}
