% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{logit_general}
\alias{logit_general}
\title{logit_general}
\usage{
logit_general(X, lb, ub)
}
\arguments{
\item{X}{scalar range [lb,ub]}

\item{lb}{lower bound for X}

\item{ub}{upper bound for X}
}
\value{
Y to be between [-inf, inf]
}
\description{
logit_general
}
\details{
bounds X which is between [lb,ub] to -inf -> inf based on the logit transformation
}
