% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_influence.R
\name{plot_influence}
\alias{plot_influence}
\title{plot_influence}
\usage{
plot_influence(influence_info, term)
}
\arguments{
\item{influence_info}{a list that has been outputed by calculate_influence() function}

\item{term}{character that represents the covariate term you want the plot to generate it for.}
}
\value{
gg plot
}
\description{
plot_influence
}
\details{
plotting function to generate CDI plots from The Bentley paper
}
