% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate_data}
\usage{
simulate_data(
  obj,
  n_sims = 100,
  fixed_effect = 0,
  random_effect = 0,
  calibrate = FALSE
)
}
\arguments{
\item{obj}{TMB object which has been optimised and checkef that it successfully converged}

\item{n_sims}{number of simulated data sets}

\item{fixed_effect}{integer[0,1] if you want to use plug-in (=0) method or Pseudo Bayes (=1) where you simulate values from the MVN joint distribution}

\item{random_effect}{integer[0,1] specifying how random effects are drawn. 0 = Bayes estiamtes, 1 = simulate a new field}

\item{calibrate}{boolean on whether we should try and calibrate the P-value.}
}
\value{
: A matrix of simulated data, with rows being each data record and cols being simulations i.e. n_sims
}
\description{
simulate_data
}
\details{
A function for simulating data from TMB model (specifically CPUEspatail obj). The idea is for the simulated data to be used for measuring consistentency with observed data using an idea analgous to prediction distributions from the Bayesian framework
This will feed into the DHARMa R package see ?DHARMa for more information
}
