% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projection.R
\name{get_projection}
\alias{get_projection}
\title{get_projection}
\usage{
get_projection(obj, data, projection_df, time_variable_label)
}
\arguments{
\item{obj}{TMB object which has been optimised and checked for convergence}

\item{data}{the tmb_data element that is returned from configure_obj() function}

\item{projection_df}{spatial data frame, should be the same one given to configure_obj()}

\item{time_variable_label}{character}
}
\value{
: projection_df with a column called predicted_y which has all the spatial components plus the time_coeffecient
}
\description{
get_projection
}
\details{
used to get the projected fitted values from a given model.
}
